# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger Tasker",
        "description": "This is a sample Petstore server.  You can find\nout more about Swagger at\n[http://swagger.io](http://swagger.io) or on\n[irc.freenode.net, #swagger](http://swagger.io/irc/).",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "apiteam@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/comments/{id}/attachments": {
            "post": {
                "tags": [
                    "Comment"
                ],
                "summary": "Upload attachment to a comment",
                "operationId": "App\\Http\\Controllers\\AttachmentController::comment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of comment",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "attachment": {
                                        "description": "file to upload",
                                        "type": "string",
                                        "format": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Return success"
                    }
                }
            }
        },
        "/tasks/{slug}/attachments": {
            "post": {
                "tags": [
                    "Task"
                ],
                "summary": "Upload attachment to a task",
                "operationId": "App\\Http\\Controllers\\AttachmentController::task",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug of task",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "attachment": {
                                        "description": "file to upload",
                                        "type": "string",
                                        "format": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Return success"
                    }
                }
            }
        },
        "/attachment/{id}": {
            "delete": {
                "tags": [
                    "Task"
                ],
                "summary": "Remove an attachment",
                "operationId": "App\\Http\\Controllers\\AttachmentController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of attachment",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return success"
                    }
                }
            }
        },
        "/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new user",
                "operationId": "addPet",
                "requestBody": {
                    "description": "Register data format",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "User' email",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "description": "User's passowrd",
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Register a user successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Log in a new user",
                "operationId": "App\\Http\\Controllers\\Auth\\AuthController::postLogin",
                "requestBody": {
                    "description": "Login data format",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "User' email",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "description": "User's passowrd",
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Register a user successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
    }
}